fix object error

+++ aiq/utils/processing.py
@@ -40,11 +40,11 @@ def neutralize(df: pd.DataFrame, industry_col: str, cap_col: str, factor_cols: l
     )
 
     # 构造回归自变量 X（行业 + 市值）
-    X = pd.concat([industry_dummies, df["feature", cap_col]], axis=1)
+    X = pd.concat([industry_dummies, df["feature", cap_col]], axis=1).astype(float)
     X = sm.add_constant(X)
 
     for factor_col in factor_cols:
-        y = df["feature", factor_col]
+        y = df["feature", factor_col].astype(float)
         neutral_factor = sm.OLS(y, X, missing="drop").fit().resid.reindex(df.index)
         df["feature", factor_col] = neutral_factor
